refactor: storeReg에서 StoreRegistration으로 폴더명 변경
