+++ src/page/StoreRegistration/view/PC/StoreRegistrationPC.module.scss
@@ -170,7 +170,7 @@
 
     &:checked {
       border: none;
-      background-image: url("../../../../assets/svg/storereg/mobile-checked.svg");
+      background-image: url("../../../../assets/svg/StoreRegistration/mobile-checked.svg");
     }
   }
